--- retrieve/params.go
@@ -1,35 +1,0 @@
-package retrieve
-
-import (
-	"github.com/filecoin-project/go-fil-markets/retrievalmarket"
-	"github.com/filecoin-project/go-fil-markets/shared"
-	"github.com/ipfs/go-cid"
-	"github.com/ipld/go-ipld-prime"
-	selectorparse "github.com/ipld/go-ipld-prime/traversal/selector/parse"
-)
-
-var dealIdGen = shared.NewTimeCounter()
-
-func RetrievalProposalForAsk(ask *retrievalmarket.QueryResponse, c cid.Cid, optionalSelector ipld.Node) (*retrievalmarket.DealProposal, error) {
-
-	if optionalSelector == nil {
-		optionalSelector = selectorparse.CommonSelector_ExploreAllRecursively
-	}
-
-	params, err := retrievalmarket.NewParamsV1(
-		ask.MinPricePerByte,
-		ask.MaxPaymentInterval,
-		ask.MaxPaymentIntervalIncrease,
-		optionalSelector,
-		nil,
-		ask.UnsealPrice,
-	)
-	if err != nil {
-		return nil, err
-	}
-	return &retrievalmarket.DealProposal{
-		PayloadCID: c,
-		ID:         retrievalmarket.DealID(dealIdGen.Next()),
-		Params:     params,
-	}, nil
-}
