+++ retrieve/retrieve.go
@@ -43,6 +43,8 @@ import (
 	"github.com/ipfs/go-graphsync/storeutil"
 	blockstore "github.com/ipfs/go-ipfs-blockstore"
 	logging "github.com/ipfs/go-log/v2"
+	"github.com/ipld/go-ipld-prime"
+	selectorparse "github.com/ipld/go-ipld-prime/traversal/selector/parse"
 	"github.com/libp2p/go-libp2p/core/host"
 	inet "github.com/libp2p/go-libp2p/core/network"
 	"github.com/libp2p/go-libp2p/core/peer"
@@ -1477,5 +1479,31 @@ func ChannelIDFromString(id string) (*datatransfer.ChannelID, error) {
 		Initiator: initiator,
 		Responder: responder,
 		ID:        datatransfer.TransferID(xferid),
+	}, nil
+}
+
+var dealIdGen = shared.NewTimeCounter()
+
+func RetrievalProposalForAsk(ask *retrievalmarket.QueryResponse, c cid.Cid, optionalSelector ipld.Node) (*retrievalmarket.DealProposal, error) {
+
+	if optionalSelector == nil {
+		optionalSelector = selectorparse.CommonSelector_ExploreAllRecursively
+	}
+
+	params, err := retrievalmarket.NewParamsV1(
+		ask.MinPricePerByte,
+		ask.MaxPaymentInterval,
+		ask.MaxPaymentIntervalIncrease,
+		optionalSelector,
+		nil,
+		ask.UnsealPrice,
+	)
+	if err != nil {
+		return nil, err
+	}
+	return &retrievalmarket.DealProposal{
+		PayloadCID: c,
+		ID:         retrievalmarket.DealID(dealIdGen.Next()),
+		Params:     params,
 	}, nil
 }
